New data/preferred-patches.clj data file, containing "preferred" patches ...
for 107 CLJ tickets that are open in the Clojure JIRA set of tickets
as of Jan 1 2013.  For 21 of them I have added :patch-extra-note keys
whose values are strings describing some extra info about the ticket
and/or patch, gleaned from my previous reports of CLJ prescreened
patches and tickets needing work.