Refine patch status and add fix{now,soon,later} messages
for patches with warnings when applied, in the patch summary reports.

The new patch status values are "warn" and "stderr", where those will
only be given if the 'git am ...' command had a 0 exit status.  "warn"
means there were warning messages sent to stderr, and stderr means
something was sent to stderr, but no warning messages.  So far for
current patches many of them have "warn" status, but I have not yet
seen any with "stderr" status.

The combination of this, and whether a patch is currently the
'preferred patch' fo the ticket, and the derived state of the ticket
is used to calculate what to do about the patch, e.g. it is ok, or it
is not OK and needs to be fixed now (derived state is past Vetted),
soon (derived state is Vetted), or later (derived state is before
Vetted).