Allow patches with apply status :warn to be considered prescreened, too.
With recent changes in the patch status, :patch-status :warn is very
close to the previous :ok, but not quite.  Typically the only
difference is that there is some trailing whitespace on some of the
lines in the patch.  Since that should be a straightforward mechanical
fix, consider such patches prescreened.