Fix #4: Move list of Clojure patch authors and email addresses out of repo
It is in my personal Dropbox account now, for easier synchronization
across multiple machines than doing it manually.

The names and email addresses are still publicly available to anyone
in Clojure JIRA attachment files, and for patches that have been
committed to Clojure, in the git commit history of Clojure, also
available to the general public.

If you want your email address to remain private, don't put it in
public.

"Three people can keep a secret, if two of them are dead."
   Benjamin Franklin